gym_bot/pyproject.toml
[project]
name = "placebo-gym"
version = "0.1.0"
description = "Telegram bot for logging gym workouts"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
